Add ability to `silence` events via `WarnErrorOptions` (#112)
* Add `silence` option to `WarnErrorOptions`

* Add `silenced` def to `IncludeExclude` and use in `includes`

* Begin using `silenced` of `WarnErrorOptions` in `warn_or_error`

Technically since we modified `includes` in the previous commit, `warn_or_error`
started including `silenced` in it's logic then. However we make it explicit
in this commit, and begin testing it.

* Refactor `get_all_subclasses` in `test_events` to ignore events classes defined in tests

In the previous commit we defined a `Note` event class based on `WarnLevel`.
The logic in `get_all_subclasses` was getting test defined event classes, which
isn't what we want. Thus here we added logic to exclude these.

* Changie doc for new `silence` of `WarnErrorOptions`

* Refactor `silence` logic from `IncludeExclude` to `WarnErrorOptions`

* Create `event_name` in `warn_or_error` to reduce regeneration of name