Refactor `EventCatcher` into a test utilities file.
We did this refactor because we plan to use the `EventCatcher` in more unit test
files. We have a similar class in dbt-core currently. It might be worthwhile at a
future date to distribute this class via the `dbt_common` package itself.

Also the empty `__init__.py` files are unfortunately necessary it seems. This is
because without them mypy will complain about "Source file found twice under
different module names".