Fix event tests to ignore event types created in testing
We had done this work previously by ignoring event classes that came
from modules beginning with `test_`. This worked because if we created
or re-implemented events in test files, their module was the the file
that they existed in. However, in the previous commit 4d349298, we
added `__init__.py` files to the directories under `tests`. This made
it so that the events that we were previously successfully ignoring
had their module updated to their full path, i.e. `tests.unit.<file_name>`,
and thus their module no longer began with `test_`, but `tests.` instead.
Because of this, we had a regression that caused the previously corrected
tests to begin failing again. This commit fixes that.