A method to add callbacks to the global event manager (#132)
* Refactor `EventCatcher` into a test utilities file.

We did this refactor because we plan to use the `EventCatcher` in more unit test
files. We have a similar class in dbt-core currently. It might be worthwhile at a
future date to distribute this class via the `dbt_common` package itself.

Also the empty `__init__.py` files are unfortunately necessary it seems. This is
because without them mypy will complain about "Source file found twice under
different module names".

* Fix event tests to ignore event types created in testing

We had done this work previously by ignoring event classes that came
from modules beginning with `test_`. This worked because if we created
or re-implemented events in test files, their module was the the file
that they existed in. However, in the previous commit 4d349298, we
added `__init__.py` files to the directories under `tests`. This made
it so that the events that we were previously successfully ignoring
had their module updated to their full path, i.e. `tests.unit.<file_name>`,
and thus their module no longer began with `test_`, but `tests.` instead.
Because of this, we had a regression that caused the previously corrected
tests to begin failing again. This commit fixes that.

* Add `add_callback` method to `IEventManager` protocol and `EventManager` class

* Add `add_callback_to_manager` for adding callbacks to global event manager

* Refactor typing of callbacks to use a central `TCallback` type definition