Backwards Compat: use `WARN_ERROR_OPTIONS` and `WARN_ERROR` from functions module as defaults
In dbt-core <= 1.9, settings for `warn_error_options` and `warn_error` is handled by
the `functions` module properties `WARN_ERROR_OPTIONS` and `WARN_ERROR` respectively.
As we don't want this upgrade of warn_or_error handling to be a breaking change, we
need to support the old way of doing things. This makes it such that the `warn_error_options`
and `warn_error` properties of the `EventManager` class will, if not set,  default to the
`function` module properties. In dbt-core 1.10+, the properties on the `EventManger` will instead
be set directly.