Backwards Compatible Version of: " Update fire_event to handle warn_or_error logic #237" (#259)
* Revert "Revert "Update `fire_event` to handle `warn_or_error` logic (#237)" (#258)"

This reverts commit 0a93135caaa8c517480d47e1a836436f1e46ac7f.

* Backwards Compat: use `WARN_ERROR_OPTIONS` and `WARN_ERROR` from functions module as defaults

In dbt-core <= 1.9, settings for `warn_error_options` and `warn_error` is handled by
the `functions` module properties `WARN_ERROR_OPTIONS` and `WARN_ERROR` respectively.
As we don't want this upgrade of warn_or_error handling to be a breaking change, we
need to support the old way of doing things. This makes it such that the `warn_error_options`
and `warn_error` properties of the `EventManager` class will, if not set,  default to the
`function` module properties. In dbt-core 1.10+, the properties on the `EventManger` will instead
be set directly.

* Back compat: ability to handle non-warning type events with warn_or_error logic

Previously if you sent a event through `warn_or_error`, no matter it's level it
would do the `warn_or_error` stuff (silencing, making it an error, etc). Of note
at present only _two_ non-warning type events are sent through `warn_or_error`. To
continue supporting this we've altered the logic for gating `warn_or_error`
to do `warn_or_error` handling IFF either the following is true:
1. `fire_event` call with `force_warn_or_error_handling=True`
2. `require_warn_or_error_handling=True` AND `event.level="warn"`

With this logic, we'll be able to move to a future where all warnings will
be eventually handled by `warn_or_error` logic by default AND other event
types can be opted in to be handled by `warn_or_error` by passing an additional
arg to `fire_event`.

* Update `TestEventManager` to support new `IEventManager` properties and `fire_event` args