Allow `exclude` to be specified for `WarnErrorOptions` when "deprecations" is present
Previously `exclude` could only be specified if `include` was set as "all" or "*". However,
with the addition of "deprecations", it now makes sense that `exclude` should be specifiable
if `include` is "all"/"*" or if "deprecations" is in either `include` or `silence`. To do
this we had to stop using `__post_init__` inherited from `IncludeExclude` as it only allows
`exclude` if `include` is "all"/"*"