`Deprecations` selection option for `WarnErrorOptions` (#266)
* Add more testing for `WarnErrorOptions`

We're about to update `WarnErrorOptions` which will further diverge it
from `IncludeExclude`. As such, we are adding more tests to ensure consisten
behavior.

* Add support for `deprecations` specification in `WarnErrorOptions`

* Allow `exclude` to be specified for `WarnErrorOptions` when "deprecations" is present

Previously `exclude` could only be specified if `include` was set as "all" or "*". However,
with the addition of "deprecations", it now makes sense that `exclude` should be specifiable
if `include` is "all"/"*" or if "deprecations" is in either `include` or `silence`. To do
this we had to stop using `__post_init__` inherited from `IncludeExclude` as it only allows
`exclude` if `include` is "all"/"*"

* Add additional comments to `WarnErrorOptions` mostly to break up code visually

* Transition WarnErrorOptions unit tests to use 'BehaviorChangeEvent'

* Add unit tests for `WarnErrorOptions` "deprecations" spec

* Add changie doc

* Change `deprecations` to `Deprecations` in `WarnErrorOptions`

* Improve/Fix comments related to `WarnErrorOptions`