Ensure `WarnErrorOptionsV2` respects pass by reference empty lists
Somtimes people do something like `warn_error_options.silence.append('my_deprecation_type')`.
By some people I mean us. We do this in the tests. When we were doing
`self.error = error or []` in `WarnErrorOptionsV2` it was breaking how this
appending worked for people using the legacy `WarnErrorOptions`, and by
people I mean us. This was because if the passed `error` is `[]` then
the statement really is `self.error = [] or []` in which case the _second_
`[]` becomes the `self.error`, breaking the ability to append to the reference.
This fixes that.