Switch the invocation context from tracking an adapter type to tracking adapters
There are _two_ conceptual changes here:

1. Tracking `adapter` instead of an `adapter_type`
This is because for our purposes we care about the _specific_ adapter
more than an _adapter type_. This is because in core we want to know
whether or not jsonschema validation should be run based on whether or
not the jsonschema supports a specific _adapter_. If we used adapter type
then we would potentially raise _incorrect_ deprecation warnings if a
third party snowflake adapter supported extra configs that we are not aware
of. By tracking the specific adapters, we will know for sure whether or now
jsonschema validations are supported for a given adapter

2. Tracking `multiple` adapters instead of an individual `adapter`
This is an instance of defensive foward thinking. There are currently
vague plans of multi adapter support. There isn't, to my knowledge, exactly
how far out on the road map, but it is a known direction. When we get to
that point, we'll need the invocation context to be able to handle knowing
about multiple adapters. Switching to an iterable now is a little odd, given
for the time being it'll only ever have one item. However, once we begin
work on multi-adapter support, this will make that easier.