Begin tracking add adapter types to invocation context (#297)
* Add adapter_type to InvocationContext

* Add changie doc

* Update the invocation context to be able to track multiple adapter types

Tracking `multiple` adapter types instead of an individual adapter type
is perhaps a premature optimization. However I'm doing it to be defensively
foward thinking. There are currently vague plans of multi adapter support.
There isn't, to my knowledge, exactly how far out on the road map, but it
is a known direction. When we get to that point, we'll need the invocation
context to be able to handle knowing about multiple adapters. Switching to
an iterable now is a little odd, given for the time being it'll only ever
have one item. However, once we begin work on multi-adapter support, this
will make that future work easier.