Move `EventCatcher` to be part of the `events` lib of the distributed package
Previously `EventCatcher` was only defined in `tests` which doesn't get distributed
as part of the `dbt-common` package. We also have an `EventCatcher` defined in
dbt-core's test suite, notably this definition has some features added to it. We also
now find ourselves needing an `EventCatcher` in dbt-adapters. As such it makes sense
to start including an `EventCatcher` as part of the dbt-common package.