Begin testing `EventCatcher` class
It feels weird to test something that has been predominantly a testing
utility. However, now that it is being included in the package, we should
test it as people could now start using it as a general utility. For
instance, during progromatic invocations of core, people could specify
an `EventCatcher` in the callbacks and then begin depending on it doing
what it supposed to do.