Create `EventType` type var, and use it to fix typing of `EventCatcher`
Using `EventCatcher` always had Mypy complaining because we'd pass in
uninstantiated events to the `event_to_catch` argument. That is we'd do
```
EventCatcher(event_to_catch=Note)
```
instead of
```
EventCatcher(event_to_catch=Note(msg=""))
```

We always did the former because it's lighter weight, and doing the
latter would require stubbing out the required event arguments. However,
the typing of the former _was_ wrong because the `EventCatcher` expected
an instantiated event, not an event type. Now we've fixed this, and can
do the former without Mypy bugging out.